[Refactoring Files] Refatorado estrutura de arquivo para modulos.
