+++ src/app.module.ts
@@ -1,23 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { CoursesModule } from './module/courses.module';
+import { CoursesModule } from './modules/course/courses.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import typeOrmAsyncConfig from './config/typeorm.config';
 
 @Module({
-  imports: [
-    CoursesModule,
-    TypeOrmModule.forRoot({
-      host: 'localhost',
-      port: 3306,
-      username: 'root',
-      password: 'root',
-      database: 'sample',
-      autoLoadEntities: true,
-      synchronize: true,
-      type: 'mysql',
-    }),
-  ],
+  imports: [TypeOrmModule.forRootAsync(typeOrmAsyncConfig), CoursesModule],
   controllers: [AppController],
   providers: [AppService],
 })
