+++ src/config/typeorm.config.ts
@@ -1,0 +1,21 @@
+import {
+  TypeOrmModuleAsyncOptions,
+  TypeOrmModuleOptions,
+} from '@nestjs/typeorm';
+
+const typeOrmAsyncConfig: TypeOrmModuleAsyncOptions = {
+  useFactory: async (): Promise<TypeOrmModuleOptions> => {
+    return {
+      type: 'mysql',
+      host: 'localhost',
+      port: 3306,
+      username: 'root',
+      password: 'root',
+      database: 'sample',
+      autoLoadEntities: true,
+      logging: true,
+    };
+  },
+};
+
+export default typeOrmAsyncConfig;
