+++ src/modules/course/services/courses.service.ts
@@ -2,9 +2,9 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { Course } from '../entities/course.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { CreateCourseDto } from '../domain/create-course.dto';
-import { UpdateCourseDto } from '../domain/update-course.dto';
-import { Tag } from '../entities/tag.entity';
+import { CreateCourseDto } from '../dto/create-course.dto';
+import { UpdateCourseDto } from '../dto/update-course.dto';
+import { Tag } from '../../tag/entities/tag.entity';
 
 @Injectable()
 export class CoursesService {
